Added authentication forms
Updating the HTML to include the sign-up, sign-in, sign-out, and
change password forms. These currently connect with and interact with
the API.

Currently, Sign Up does not work as it requires additional inputs to
create a new user in the API (the equipped_item slots). This will either
be solved by making those columns non-mandatory for user creation or
updating the forms to include this information.