Added hide/show and colors to app's display
Adding color-coded messaging (green for success, red for failure, grey
for other) for app functions, liking signing in and creating loadouts.
Additionally, actions like Sign-In and Sign-Out now hide and show
app forms and functions depending on if they apply to the current
user's login state.

e.g. 'Create Loadout' doesn't show until a user  signs in, since a user
needs to be signed in to create a loadout. Similarly, the 'Sign In' form
is hidden after signing in, and remains hidden until a user signs out.