New Character Section and moved "Show" function
Updating the app to now have a new character block that includes the
currently-used loadout and the currently-equipped armor, weapon, and
trinket.

The "Show" button is now located on the indexed list of loadouts.
Clicking it not sets the current loadout (and all its properties) as the
currently equipped ones on the character.

Removed the previous "Show" button as it was no longer useful.