Added "New Loadout" form and updated design
Adding a new loadout form to separate the resource creating and resource
patching processes. Previously, they were to be located on the same
form, but that provided challenges. This can be reattempted later.

Additionally, the app received substantial design changes. This includes
formatting, coloring, and positioning of different sections of the app.
Little changed in the way of scripts with the design changes.